split out all remote indexing functionality into isolated zomes for clean separation of entry & link type definitions
