+++ Cargo.toml
@@ -58,10 +58,12 @@ members= [
   "happs/observation/zomes/fulfillment/code",
   "happs/observation/zomes/satisfaction/code",
   "happs/observation/zomes/process/code",
+  "happs/observation/zomes/resource_specification/code",
   "happs/planning/zomes/commitment/code",
   "happs/planning/zomes/fulfillment/code",
   "happs/planning/zomes/intent/code",
   "happs/planning/zomes/satisfaction/code",
+  "happs/planning/zomes/process/code",
   "happs/specification/zomes/resource_specification/code",
   "happs/specification/zomes/process_specification/code",
   "happs/specification/zomes/unit/code",
