+++ happs/observation/zomes/resource_specification/code/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "hc_zome_rea_resource_specification_index_observation"
+version = "0.1.0"
+authors = ["pospi <pospi@spadgos.com>"]
+edition = "2018"
+
+[dependencies]
+serde = "=1.0.89"
+serde_json = { version = "=1.0.39", features = ["preserve_order"] }
+# :DUPE: hdk-rust-revid
+hdk = "=0.0.42-alpha5"
+hdk_proc_macros = "=0.0.42-alpha5"
+
+hc_zome_rea_resource_specification_storage_consts = { path = "../../../../../lib/rea_resource_specification/storage_consts" }
+hc_zome_rea_economic_resource_storage_consts = { path = "../../../../../lib/rea_economic_resource/storage_consts" }
+
+[lib]
+path = "src/lib.rs"
+crate-type = ["cdylib"]
