+++ happs/observation/zomes/resource_specification/zome.json
@@ -1,0 +1,3 @@
+{
+  "description": "Holo-REA resource specification indexing functionality for the Holo-REA &#39;observation&#39; DNA."
+}
