+++ happs/planning/zomes/commitment/code/Cargo.toml
@@ -15,9 +15,6 @@ hdk_proc_macros = "0.0.42-alpha5"
 hc_zome_rea_commitment_defs = { path = "../../../../../lib/rea_commitment/defs" }
 hc_zome_rea_commitment_rpc = { path = "../../../../../lib/rea_commitment/rpc" }
 hc_zome_rea_commitment_lib = { path = "../../../../../lib/rea_commitment/lib" }
-hc_zome_rea_commitment_storage_consts = { path = "../../../../../lib/rea_commitment/storage_consts" }
-hc_zome_rea_intent_storage_consts = { path = "../../../../../lib/rea_intent/storage_consts" }
-hc_zome_rea_process_storage_consts = { path = "../../../../../lib/rea_process/storage_consts" }
 
 [lib]
 path = "src/lib.rs"
