+++ happs/planning/zomes/process/zome.json
@@ -1,0 +1,3 @@
+{
+  "description": "Holo-REA process indexing functionality for the Holo-REA &#39;planning&#39; DNA."
+}
